8-Attributes Management part 3

+++ database/migrations/2020_03_26_220133_create_attribute_values_table.php
@@ -18,7 +18,8 @@ class CreateAttributeValuesTable extends Migration
             $table->unsignedBigInteger('attribute_id')->unsigned();
             $table->foreign('attribute_id')->references('id')->on('attributes');
             $table->text('value');
-            $table->decimal('price', 2)->nullable();
+//            $table->decimal('price', 2)->nullable();
+            $table->double('price', 2)->nullable();
             $table->timestamps();
         });
     }
